Refactor vendor document group controller to use service layer
